Split trigger class into two partial halves for the 2 endpoints
